refactor(flame_bloc): rename FlameBlocListener component and mixin (#1599)
