+++ packages/flame_bloc/lib/flame_bloc.dart
@@ -1,7 +1,7 @@
 export 'src/flame_bloc_game.dart';
 
+export 'src/flame_bloc_listenable.dart';
 export 'src/flame_bloc_listener.dart';
-export 'src/flame_bloc_listener_component.dart';
 export 'src/flame_bloc_provider.dart';
 export 'src/flame_bloc_reader.dart';
 export 'src/flame_multi_bloc_provider.dart';
